Refactor using NextJS params matching
